python/django-opensearch/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
dj-database-url==3.0.1
Django==5.2.6
django-elasticsearch-dsl==9.0
django-seed==0.3.1
djangorestframework==3.16.1
dotenv==0.9.9
elastic-transport==9.1.0
elasticsearch==9.1.1
Faker==37.8.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
